Code reorganised: import.py renamed to app_import.py
*** All GUI elements in all classes in dedicated function
deleted mainloop() from all classes except main gui
