+++ .gitignore
@@ -6,4 +6,6 @@
 !root_app.ui
 !root_login.ui
 !addTransaction.py
-!import.py
+!app_import.py
+!app_login.py
+
