+++ AppGui.py
@@ -14,6 +14,7 @@ from app_data import App_data
 from addTransaction import AddTransaction
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 import matplotlib.pyplot as plt
+import app_import
 ### ***
 #import lotto
 
@@ -23,202 +24,6 @@ _cal_datefmt = "dd.mm.yyyy"
 
 #######################################3
 
-class LoginWin:
-    def __init__(self, master=None):
-        
-        # build ui
-        if master == None:
-            self.root_login = tk.Tk()
-        else:
-            self.root_login = tk.Toplevel(master)
-            
-        ## Hide window 
-        ## DO NOT forget to show at the end of init!!!
-        self.root_login.withdraw()     
-            
-        self.ntb_Login = ttk.Notebook(self.root_login)
-        
-        #LOG IN TAB
-        self.frm_login = ttk.Frame(self.ntb_Login)
-        self.lbfr_login = ttk.Labelframe(self.frm_login)
-        self.lbl_selectUser = ttk.Label(self.lbfr_login)
-        self.lbl_selectUser.configure(text='Select user')
-        self.lbl_selectUser.grid(column='0', row='0', sticky='s')
-        self.lbl_selectUser.master.rowconfigure('0', pad='20', weight='0')
-        self.lbl_selectUser.master.columnconfigure('0', pad='0', weight='1')
-        self.cmb_UserLogin = ttk.Combobox(self.lbfr_login, 
-                                          state="readonly")
-        self.cmb_UserLogin.configure(width='20')
-        self.cmb_UserLogin.grid(column='0', pady='10', row='1')
-        self.cmb_UserLogin.master.columnconfigure('0', pad='0', weight='1')
-        self.lbl_PasswordLogin = ttk.Label(self.lbfr_login)
-        self.lbl_PasswordLogin.configure(text='Enter password')
-        self.lbl_PasswordLogin.grid(column='0', row='2', sticky='s')
-        self.lbl_PasswordLogin.master.rowconfigure('2', pad='15')
-        self.lbl_PasswordLogin.master.columnconfigure('0', pad='0', weight='1')
-        self.txt_PasswordLogin = ttk.Entry(self.lbfr_login)
-        self.txt_PasswordLogin.configure(show='*', width='20')
-        self.txt_PasswordLogin.grid(column='0', pady='10', row='3')
-        self.txt_PasswordLogin.master.columnconfigure('0', pad='0', weight='1')
-        self.btn_Login = ttk.Button(self.lbfr_login, 
-                                    command = self.cb_Login)
-        self.btn_Login.configure(text='LOG IN', width='15')
-        self.btn_Login.grid(column='0', pady='20', row='4')
-        self.btn_Login.master.columnconfigure('0', pad='0', weight='1')
-        self.lbfr_login.configure(height='0', text='Please log in below', width='0')
-        self.lbfr_login.grid(column='0', ipady='30', padx='40', pady='40', row='0', sticky='nsew')
-        self.lbfr_login.master.rowconfigure('0', pad='10', weight='1')
-        self.lbfr_login.master.columnconfigure('0', pad='20', weight='1')
-        self.frm_login.configure(height='0', width='0')
-        self.frm_login.grid(column='0', row='0', sticky='nsew')
-        self.frm_login.master.rowconfigure('0', weight='1')
-        self.frm_login.master.columnconfigure('0', weight='1')
-        self.ntb_Login.add(self.frm_login, state='hidden', sticky='nsew', text='log in')        
-        
-        
-        # ADD USER TAB
-        self.frm_addUser = ttk.Frame(self.ntb_Login)
-        self.lbfr_addUser = ttk.Labelframe(self.frm_addUser)
-        self.lbl_InfoAddUser = ttk.Label(self.lbfr_addUser)
-        self.lbl_InfoAddUser.configure(justify='center', text='Please, enter new user name\n and type your password two times')
-        self.lbl_InfoAddUser.grid(column='0', columnspan='2', pady='15', row='0')
-        self.lbl_InfoAddUser.master.rowconfigure('0', pad='0')
-        self.lbl_InfoAddUser.master.columnconfigure('0', pad='10')
-        self.lbl_usernameAddUser = ttk.Label(self.lbfr_addUser)
-        self.lbl_usernameAddUser.configure(text='Username')
-        self.lbl_usernameAddUser.grid(column='0', padx='11', pady='10', row='1', sticky='e')
-        self.lbl_usernameAddUser.master.rowconfigure('1', pad='10')
-        self.lbl_usernameAddUser.master.columnconfigure('0', pad='10')
-        self.txt_usernameAddUser = ttk.Entry(self.lbfr_addUser)
-        self.txt_usernameAddUser.grid(column='1', row='1', sticky='w')
-        self.txt_usernameAddUser.master.rowconfigure('1', pad='10')
-        self.txt_usernameAddUser.master.columnconfigure('1', pad='11', weight='1')
-        self.lbl_PasswordAdd = ttk.Label(self.lbfr_addUser)
-        self.lbl_PasswordAdd.configure(text='Password')
-        self.lbl_PasswordAdd.grid(column='0', padx='11', pady='5', row='2', sticky='e')
-        self.lbl_PasswordAdd.master.rowconfigure('2', pad='5')
-        self.lbl_PasswordAdd.master.columnconfigure('0', pad='10')
-        self.txt_passwordAdd = ttk.Entry(self.lbfr_addUser)
-        self.txt_passwordAdd.configure(show='*')
-        self.txt_passwordAdd.grid(column='1', row='2', sticky='w')
-        self.txt_passwordAdd.master.rowconfigure('2', pad='5')
-        self.txt_passwordAdd.master.columnconfigure('1', pad='11', weight='1')
-        self.lbl_PasswordConfAdd = ttk.Label(self.lbfr_addUser)
-        self.lbl_PasswordConfAdd.configure(text='Confirm password')
-        self.lbl_PasswordConfAdd.grid(column='0', padx='11', pady='5', row='3', sticky='e')
-        self.lbl_PasswordConfAdd.master.rowconfigure('3', pad='5')
-        self.lbl_PasswordConfAdd.master.columnconfigure('0', pad='10')
-        self.txt_passwordConfAdd = ttk.Entry(self.lbfr_addUser)
-        self.txt_passwordConfAdd.configure(show='*')
-        self.txt_passwordConfAdd.grid(column='1', row='3', sticky='w')
-        self.txt_passwordConfAdd.master.rowconfigure('3', pad='5')
-        self.txt_passwordConfAdd.master.columnconfigure('1', pad='11', weight='1')
-        self.btn_submitAdd = ttk.Button(self.lbfr_addUser, 
-                                        command = self.cb_CreateUser)
-        self.btn_submitAdd.configure(text='SUBMIT', width='15')
-        self.btn_submitAdd.grid(column='0', columnspan='2', ipadx='10', ipady='5', padx='10', pady='15', row='4')
-        self.btn_submitAdd.master.columnconfigure('0', pad='10')
-        self.lbl_resultAdduser = ttk.Label(self.lbfr_addUser)
-        self.var_AddUserResult = tk.StringVar(value='Enter username, password and press Submit button')
-        self.lbl_resultAdduser.configure(textvariable=self.var_AddUserResult)
-        self.lbl_resultAdduser.grid(column='0', columnspan='2', row='5')
-        self.lbl_resultAdduser.master.columnconfigure('0', pad='10')
-        self.lbfr_addUser.configure(height='0', text='Please enter the details below', width='0')
-        self.lbfr_addUser.grid(column='0', ipady='30', padx='40', pady='40', row='0', sticky='nsew')
-        self.lbfr_addUser.master.rowconfigure('0', weight='1')
-        self.lbfr_addUser.master.columnconfigure('0', weight='1')
-        self.frm_addUser.configure(height='0', width='0')
-        self.frm_addUser.grid(column='0', row='0', sticky='nsew')
-        self.frm_addUser.master.rowconfigure('0', weight='1')
-        self.frm_addUser.master.columnconfigure('0', weight='1')
-        self.ntb_Login.add(self.frm_addUser, sticky='nsew', text='add user')
-        
-        ## Window settings
-        self.ntb_Login.configure(width='400')
-        self.ntb_Login.grid(column='0', padx='3', pady='3', row='0', sticky='nsew')
-        self.ntb_Login.master.rowconfigure('0', minsize='0', weight='1')
-        self.ntb_Login.master.columnconfigure('0', minsize='0', weight='1')
-        self.root_login.configure(height='150', takefocus=False, width='400')
-        self.root_login.resizable(False, False)
-
-        ## Style correcting for tab's well-view
-        style = ttk.Style()                     
-        current_theme =style.theme_use()
-        style.theme_settings(current_theme, 
-                {"TNotebook.Tab": {"configure": {"padding": [95, 5]}}})
-
-        # SHOW window, fully constructed
-        self.root_login.deiconify()
-    
-        # Main widget
-        self.loginWindow = self.root_login
-        self.loginWindow.title("Login")
-
-    def run(self):
-        self.loginWindow.grab_set()
-        self.loginWindow.mainloop()
-
-
-    def cb_CreateUser(self):
-        """
-        Create user callback
-
-        Returns
-        -------
-        None.
-
-        """
-
-        if not self.txt_usernameAddUser.get():
-            self.lbl_usernameAddUser.config(foreground="red")
-            self.var_AddUserResult.set(value="User name cannot be empty")
-            return False
-        else:
-            self.lbl_usernameAddUser.configure(foreground="black")
-        
-        if not self.txt_passwordAdd.get():
-            self.lbl_PasswordAdd.config(foreground="red")
-            self.var_AddUserResult.set(value="Password cannot be empty")
-            return False
-        else:
-            self.lbl_PasswordAdd.configure(foreground="black")
-        
-        if not (self.txt_passwordAdd.get() == self.txt_passwordConfAdd.get()):
-            self.lbl_PasswordAdd.config(foreground="red")
-            self.lbl_PasswordConfAdd.config(foreground="red")
-            self.var_AddUserResult.set(value="Passwords are not match")
-            return False
-        else:
-            self.lbl_PasswordAdd.configure(foreground="black")
-            self.lbl_PasswordConfAdd.configure(foreground="black")
-        ## TODO: call UserCreate()
-        self.var_AddUserResult.set(value="Creating user...")
-        pass
-    
-    
-    def cb_Login(self):
-        """
-        Create user callback
-
-        Returns
-        -------
-        None.
-
-        """
-        if  self.cmb_UserLogin.current() < 0:
-            self.lbl_selectUser.config(foreground="red")
-            return False
-        else:
-            self.lbl_selectUser.configure(foreground="black")
-
-        if not self.txt_PasswordLogin.get():
-            self.lbl_PasswordLogin.config(foreground="red")
-            return False
-        else:
-            self.lbl_PasswordLogin.configure(foreground="black")
-        
-        ## TODO: call UserLogin() 
-        pass
 
 class AppWin:
     def __init__(self, master=None):
@@ -226,8 +31,205 @@ class AppWin:
         # DataHandler instance
         self.badb = App_data()
         
+        # TODO: user manager handle
+        
+        # TODO: check user and call loginDialog ???
+        
+        # Main widget, build GUI
+        self.mainwindow = self.GUI(master)
+        ### BINDs
+        self.cal_tr_To.bind('<<DateEntrySelected>>', lambda x: self.updateTransactionTable() )
+        self.cal_tr_From.bind('<<DateEntrySelected>>', lambda x: self.updateTransactionTable() )
+        
+        
+
+
+        
+    def chartSpendingsMonth(self):
+        mon = self.spn_month.get()
+        
+        def getmonth():
+            if mon == "January": return "01"
+            elif mon =="February": return "02"
+            elif mon == "March": return "03"
+            elif mon == "April": return "04"
+            elif mon == "May": return "05"
+            elif mon == "June": return "06"
+            elif mon == "July": return "07"
+            elif mon == "August": return "08"
+            elif mon == "September": return "09"
+            elif mon == "October": return "10"
+            elif mon == "December": return "12"
+            elif mon == "November": return "11"
+            else: 
+              tk.messagebox.showwarning("Spinbox!!!!","put correct month",
+                                      parent=self.mainwindow)
+               
+        monthname = str(mon)
+        month=str(getmonth())
+        amount = 0
+        category = 1
+        
+        am = [-i[amount] for i in self.badb.chartMonth(month)]
+        cat=[i[category] for i in self.badb.chartMonth(month)]
+        print(i for i in self.badb.chartMonth(month))
+    
+        #fig = plt.figure(dpi=dpi)
+        fig = plt.figure(dpi=100)
+        ax = fig.add_subplot(111)
+        chart = FigureCanvasTkAgg(fig, self.lbfr_Acc_Chart)
+        chart.get_tk_widget().grid(padx=5, pady=5,
+                                         column="0",row="1",columnspan="2")
+        ax.bar(cat,height=am)
+        ax.set_title('Spendings in '+ monthname)
+        ax.set_xlabel("Categories");ax.set_ylabel("Spendings [Euros]")
+
+
+        
+    def updateTransactionTable(self):
+        # self.display_balance()
+        #first clear the treeview
+        for i in self.tbl_transactions.get_children():
+            self.tbl_transactions.delete(i)
+             
+        #then display data
+        datefr = self.cal_tr_From.get_date()
+        dateto = self.cal_tr_To.get_date()
+        
+        data = self.badb.getAllTransactionsPeriod(datefr, dateto)
+        for row in data:
+            idvalue = row[0]
+            date = row[1].strftime(_dt_datefmt)
+            cat = row[3] if row[3] else ""
+            con = row[4] if row[4] else ""
+            values = (idvalue,date, row[2], cat, con)
+            self.tbl_transactions.insert('','end', values = values)
+
+
+
+    def updateCategoriesTable(self):
+        #first clear the treeview
+        for item in  self.tbl_categories.get_children():
+             self.tbl_categories.delete(item)
+        #then display data
+        data = self.badb.getCategoriesList()
+        for row in data:
+            cat = row[0] if row[0] else ""
+            _id = row[1] if row[1] else ""
+            values = (_id, cat)
+            self.tbl_categories.insert('','end', values = values)
+        
+
+
+    def h_btnTrAdd(self):
+        addTransactionWindow = AddTransaction(self.mainwindow, self)
+        #self.mainwindow.wait_window(addTransactionWindow)
+        self.updateTransactionTable()
+
+
+    def h_btnLogout(self):
+        pass
+
+
+    def treeSelection(self):
+        line = self.tbl_transactions.selection()
+        value = self.tbl_transactions.item(line)['values']
+        return value
+        
+        
+    def h_btnTrChange(self):
+        try:
+            id_value = str(self.treeSelection()[0])
+        except:
+            print("Nothing selected in table. Cannot change.")
+            tk.messagebox.showinfo("Change transaction","Select transaction in table to change.",
+                                      parent=self.mainwindow)
+            return
+        print(id_value)
+        AddTransaction(self.mainwindow, self, id_value)
+        self.updateTransactionTable()
+        
+
+    def h_btnTrDelete(self):
+        try:
+            id_value = str(self.treeSelection()[0])
+        except:
+            print("Nothing selected. Cannot delete.")
+            return
+        reply =  tk.messagebox.askyesno(title="Delete...", 
+                       message="You going to delete transaction.\n\n" + 
+                               "Are you sure?")
+        if reply == True:
+            self.badb.deleteTransaction(id_value)
+            self.updateTransactionTable()
+
+
+    def h_btnTrImport(self):
+        importWindow = app_import.ImportTransactionDialog(self.mainwindow, self)
+        importWindow.run()
+        print("Import dialog is opened. waiting...")
+        self.mainwindow.wait_window(importWindow.mainwindow)
+        print(".....end wait. Import dialog closed.")
+        self.updateTransactionTable()
+        pass
+
+
+    def h_btnTrExport(self):
+        pass
+
+
+    def h_btnTrLotto(self):
+        date = dt.date.today()
+        amount = lotto.check()
+        self.badb.addTransaction(date, amount, "Lotto", None)
+        self.updateTransactionTable()
+        if amount < 0:
+            tk.messagebox.showwarning("LOTTO RESULTS!!!!","You have bad luck",
+                                      parent=self.mainwindow)
+        else:
+            tk.messagebox.showwarning("LOTTO RESULTS!!!!","You won "+str(amount)+", yeay",
+                                      parent=self.mainwindow)
+
+
+    def h_btnCatAdd(self):
+        pass
+
+
+    def h_btnCatChange(self):
+        pass
+
+
+    def h_btnCatDelete(self):
+        pass
+
+               
+        
+    def display_time(self):
+        self.var_CurrentTime.set( value= time.strftime('%H:%M:%S') )
+        self.mainwindow.after(1000, self.display_time)     
+
+    def display_balance(self):
+        ## TODO: not by after
+        val = self.badb.getBalance()
+        self.var_CurrentBalance.set( value= f"{val:.2f}" )
+    
+
+    def run(self):
+        self.treeSelection()
+        self.updateTransactionTable()
+        self.updateCategoriesTable()
+        self.display_time()
+        self.display_balance()
+        self.mainwindow.mainloop()    
+
+
+    def GUI(self, master):
         # build ui
-        self.root_app = tk.Tk() if master is None else tk.Toplevel(master)
+        if master == None:
+            self.root_app = tk.Tk()
+        else:
+            self.root_app = tk.Toplevel(master)
+        
         ## Hide window 
         ## DO NOT forget to show at the end of init!!!
         self.root_app.withdraw()     
@@ -323,11 +325,11 @@ class AppWin:
         self.label1.configure(text='Category:')
         self.label1.grid(column='0', padx='10', pady='10', row='2', sticky='e')
         self.label1.master.rowconfigure('1', pad='10')
-        self.label1.master.rowconfigure('2', pad='10')
+        self.label1.master.rowconfigure('2', pad='5')
         self.label1.master.columnconfigure('0', pad='10')
         self.cmb_tr_Category = ttk.Combobox(self.lbfr_drTransactions)
         self.cmb_tr_Category.grid(column='1', row='2', sticky='ew')
-        self.lbfr_drTransactions.configure(height='0', text='Choose date range')
+        self.lbfr_drTransactions.configure(height='0', padding='0 0 20 0', text='Choose date range')  #
         self.lbfr_drTransactions.grid(column='0', ipadx='0', ipady='0', padx='5', pady='0', row='0', sticky='nsew')
         self.lbfr_drTransactions.master.rowconfigure('0', pad='10', weight=0)
         self.lbfr_drTransactions.master.columnconfigure('0', pad='0', weight=1)
@@ -532,187 +534,12 @@ class AppWin:
         self.root_app.resizable(True, True)
         self.root_app.title('Python cash')
         
-        ### BINDs
-        self.cal_tr_To.bind('<<DateEntrySelected>>', lambda x: self.updateTransactionTable() )
-        self.cal_tr_From.bind('<<DateEntrySelected>>', lambda x: self.updateTransactionTable() )
-        
         # SHOW window, fully constructed
         self.root_app.deiconify()
-
-        # Main widget
-        self.mainwindow = self.root_app
-
-
-        
-    def chartSpendingsMonth(self):
-        mon = self.spn_month.get()
-        
-        def getmonth():
-            if mon == "January": return "01"
-            elif mon =="February": return "02"
-            elif mon == "March": return "03"
-            elif mon == "April": return "04"
-            elif mon == "May": return "05"
-            elif mon == "June": return "06"
-            elif mon == "July": return "07"
-            elif mon == "August": return "08"
-            elif mon == "September": return "09"
-            elif mon == "October": return "10"
-            elif mon == "December": return "12"
-            elif mon == "November": return "11"
-            else: 
-              tk.messagebox.showwarning("Spinbox!!!!","put correct month",
-                                      parent=self.mainwindow)
-               
-        monthname = str(mon)
-        month=str(getmonth())
-        amount = 0
-        category = 1
-        
-        am = [-i[amount] for i in self.badb.chartMonth(month)]
-        cat=[i[category] for i in self.badb.chartMonth(month)]
-        print(i for i in self.badb.chartMonth(month))
-    
-        #fig = plt.figure(dpi=dpi)
-        fig = plt.figure(dpi=100)
-        ax = fig.add_subplot(111)
-        chart = FigureCanvasTkAgg(fig, self.lbfr_Acc_Chart)
-        chart.get_tk_widget().grid(padx=5, pady=5,
-                                         column="0",row="1",columnspan="2")
-        ax.bar(cat,height=am)
-        ax.set_title('Spendings in '+ monthname)
-        ax.set_xlabel("Categories");ax.set_ylabel("Spendings [Euros]")
-
-
-        
-    def updateTransactionTable(self):
-        # self.display_balance()
-        #first clear the treeview
-        self.tbl_transactions.delete(*self.tbl_transactions.get_children())
-             
-        #then display data
-        datefr = self.cal_tr_From.get_date()
-        dateto = self.cal_tr_To.get_date()
-        
-        data = self.badb.getAllTransactionsPeriod(datefr, dateto)
-        for row in data:
-            idvalue = row[0]
-            date = row[1].strftime(_dt_datefmt)
-            cat = row[3] if row[3] else ""
-            con = row[4] if row[4] else ""
-            values = (idvalue,date, row[2], cat, con)
-            self.tbl_transactions.insert('','end', values = values)
-
-
-
-    def updateCategoriesTable(self):
-        #first clear the treeview
-        for item in  self.tbl_categories.get_children():
-             self.tbl_categories.delete(item)
-        #then display data
-        data = self.badb.getCategoriesList()
-        for row in data:
-            cat = row[0] if row[0] else ""
-            _id = row[1] if row[1] else ""
-            values = (_id, cat)
-            self.tbl_categories.insert('','end', values = values)
-        
-
-
-    def h_btnTrAdd(self):
-        addTransactionWindow = AddTransaction(self.mainwindow, self)
-        addTransactionWindow.run()
-        #self.mainwindow.wait_window(addTransactionWindow)
-        #self.updateTransactionTable()
-
-
-    def h_btnLogout(self):
-        pass
-
-
-    def treeSelection(self):
-        line = self.tbl_transactions.selection()
-        value = self.tbl_transactions.item(line)['values']
-        return value
-        
-        
-    def h_btnTrChange(self):
-        id_value = str(self.treeSelection()[0])
-        print(id_value)
-        AddTransaction(self.mainwindow, self, id_value).run()
-        self.updateTransactionTable()
-        
-
-    def h_btnTrDelete(self):
-        try:
-            id_value = str(self.treeSelection()[0])
-        except:
-            print("Nothing selected. Cannot delete.")
-            return
-        reply =  tk.messagebox.askyesno(title="Delete...", 
-                       message="You going to delete transaction.\n\n" + 
-                               "Are you sure?")
-        if reply == True:
-            self.badb.deleteTransaction(id_value)
-            self.updateTransactionTable()
-
-
-    def h_btnTrImport(self):
-        pass
-
-
-    def h_btnTrExport(self):
-        pass
-
-
-    def h_btnTrLotto(self):
-        date = dt.date.today()
-        amount = lotto.check()
-        self.badb.addTransaction(date, amount, "Lotto", None)
-        self.updateTransactionTable()
-        if amount < 0:
-            tk.messagebox.showwarning("LOTTO RESULTS!!!!","You have bad luck",
-                                      parent=self.mainwindow)
-        else:
-            tk.messagebox.showwarning("LOTTO RESULTS!!!!","You won "+str(amount)+", yeay",
-                                      parent=self.mainwindow)
-
-
-    def h_btnCatAdd(self):
-        pass
-
-
-    def h_btnCatChange(self):
-        pass
-
-
-    def h_btnCatDelete(self):
-        pass
-
-               
-        
-    def display_time(self):
-        self.var_CurrentTime.set( value= time.strftime('%H:%M:%S') )
-        self.mainwindow.after(1000, self.display_time)     
-
-    def display_balance(self):
-        ## TODO: not by after
-        val = self.badb.getBalance()
-        self.var_CurrentBalance.set( value= f"{val:.2f}" )
-    
-
-    def run(self):
-        self.treeSelection()
-        self.updateTransactionTable()
-        self.updateCategoriesTable()
-        self.display_time()
-        self.display_balance()
-        self.mainwindow.mainloop()    
-
+        ### 
+        return self.root_app
 
 if __name__ == '__main__':
-    #app = LoginWin()
-    #app.run()
     
     app = AppWin()
     app.run()
